+++ src/Config/HttpRequest.php
@@ -2,7 +2,7 @@
 
 namespace MedyaT\Parapos\Config;
 
-use MedyaT\Parapos\Models\Payment;
+use Illuminate\Database\Eloquent\Model;
 
 final class HttpRequest implements \Stringable
 {
@@ -10,7 +10,7 @@ final class HttpRequest implements \Stringable
      * @param  string[]  $headers
      * @param  mixed[]  $params
      */
-    public function __construct(public Payment $payment, public string $uri, public string $method, public array $headers, public array $params = []) {}
+    public function __construct(public Model $payment, public string $uri, public string $method, public array $headers, public array $params = []) {}
 
     public function __toString(): string
     {